# website/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Python Demo API",
    "summary": "A compact FastAPI reference app with rich logging and a Books API.",
    "description": "Companion demo app for the concepts discussed in the docs. The live API focuses on request validation, pagination, partial updates, rich logging, health checks, redirects, and safe demo controls. Authentication and cloud deployment patterns are explained in the documentation.",
    "version": "0.3.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [
          "Monitoring"
        ],
        "summary": "Root",
        "description": "Return high-level API metadata and local documentation links.",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RootMetadata"
                }
              }
            }
          }
        }
      }
    },
    "/api/ps": {
      "get": {
        "tags": [
          "Monitoring"
        ],
        "summary": "Ps",
        "description": "Return a lightweight health payload for local smoke testing.",
        "operationId": "ps_api_ps_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/status": {
      "head": {
        "tags": [
          "Monitoring"
        ],
        "summary": "Status Head",
        "description": "Return a header-only status response for monitoring checks.",
        "operationId": "status_head_api_status_head",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/options": {
      "options": {
        "tags": [
          "HTTP Behaviors"
        ],
        "summary": "Options Test",
        "description": "Return supported methods for quick HTTP tooling demos.",
        "operationId": "options_test_api_options_options",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/redirect": {
      "get": {
        "tags": [
          "HTTP Behaviors"
        ],
        "summary": "Redirect Test",
        "description": "Redirect to a random seeded book to demonstrate 307 behavior.",
        "operationId": "redirect_test_api_redirect_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/books": {
      "get": {
        "tags": [
          "Books API"
        ],
        "summary": "List Books",
        "description": "List books with offset pagination.",
        "operationId": "list_books_api_v1_books_get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Books API"
        ],
        "summary": "Create Book",
        "description": "Create a book and add it to the in-memory store.",
        "operationId": "create_book_api_v1_books_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/books/search": {
      "get": {
        "tags": [
          "Books API"
        ],
        "summary": "Search Books",
        "description": "Search books by title or author substring.",
        "operationId": "search_books_api_v1_books_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "title": "Q"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/books/{book_id}": {
      "get": {
        "tags": [
          "Books API"
        ],
        "summary": "Get Book",
        "description": "Return a single book by UUID.",
        "operationId": "get_book_api_v1_books__book_id__get",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Book Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Books API"
        ],
        "summary": "Update Book",
        "description": "Apply a partial update to a book.",
        "operationId": "update_book_api_v1_books__book_id__patch",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Book Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Books API"
        ],
        "summary": "Delete Book",
        "description": "Delete a book from the in-memory store.",
        "operationId": "delete_book_api_v1_books__book_id__delete",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Book Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Delete Book Api V1 Books  Book Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BookCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "author": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Author"
          },
          "year": {
            "type": "integer",
            "maximum": 2100.0,
            "minimum": 1900.0,
            "title": "Year"
          }
        },
        "type": "object",
        "required": [
          "title",
          "author",
          "year"
        ],
        "title": "BookCreate"
      },
      "BookListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BookResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "skip": {
            "type": "integer",
            "title": "Skip"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "skip",
          "limit"
        ],
        "title": "BookListResponse"
      },
      "BookResponse": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "author": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Author"
          },
          "year": {
            "type": "integer",
            "maximum": 2100.0,
            "minimum": 1900.0,
            "title": "Year"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "title",
          "author",
          "year",
          "id",
          "created_at"
        ],
        "title": "BookResponse"
      },
      "BookUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "author": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2100.0,
                "minimum": 1900.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "type": "object",
        "title": "BookUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "books": {
            "type": "integer",
            "title": "Books"
          },
          "traffic_generator_enabled": {
            "type": "boolean",
            "title": "Traffic Generator Enabled"
          },
          "chaos_headers_enabled": {
            "type": "boolean",
            "title": "Chaos Headers Enabled"
          }
        },
        "type": "object",
        "required": [
          "status",
          "books",
          "traffic_generator_enabled",
          "chaos_headers_enabled"
        ],
        "title": "HealthResponse"
      },
      "RootMetadata": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "docs_url": {
            "type": "string",
            "title": "Docs Url"
          },
          "redoc_url": {
            "type": "string",
            "title": "Redoc Url"
          },
          "openapi_url": {
            "type": "string",
            "title": "Openapi Url"
          },
          "traffic_generator_enabled": {
            "type": "boolean",
            "title": "Traffic Generator Enabled"
          },
          "chaos_headers_enabled": {
            "type": "boolean",
            "title": "Chaos Headers Enabled"
          }
        },
        "type": "object",
        "required": [
          "name",
          "version",
          "docs_url",
          "redoc_url",
          "openapi_url",
          "traffic_generator_enabled",
          "chaos_headers_enabled"
        ],
        "title": "RootMetadata"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "tags": [
    {
      "name": "Books API",
      "description": "CRUD, pagination, search, and partial updates for the demo resource."
    },
    {
      "name": "Monitoring",
      "description": "Health and status endpoints used for smoke tests and uptime checks."
    },
    {
      "name": "HTTP Behaviors",
      "description": "Small endpoints that demonstrate redirects and method negotiation."
    }
  ]
}